requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
flake8==3.8.3
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
install==1.3.3
isort==5.2.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0